+++ SourcesAXR/xrGameCS/WeaponMagazinedWGrenade.h
@@ -86,7 +86,7 @@ private:
 	virtual bool	install_upgrade_impl		( LPCSTR section, bool test );
 	virtual	bool	install_upgrade_ammo_class	( LPCSTR section, bool test );
 
-			int		GetAmmoCount2				( u8 ammo2_type ) const;
+			int		GetAmmoCount2				(u32 ammo2_type ) const;
 	
 public:
 	//дополнительные параметры патронов
