+++ SourcesAXR/xrGameSoC/ui/UIBoosterInfo.cpp
@@ -339,8 +339,8 @@ void CUIBoosterInfo::SetInfo(CInventoryItem& pInvItem)
 	//Portions
 	if (eatable)
 	{
-		val = eatable->GetPortionsNum();
-		max_value = eatable->GetConstPortionsNum();
+		val = (float)eatable->GetPortionsNum();
+		max_value = (float)eatable->GetConstPortionsNum();
 
 		if (max_value > 1)
 		{
