+++ SourcesAXR/xrGameSoC/ui/UIInventoryItemParams.cpp
@@ -377,7 +377,7 @@ void CUIInventoryItem::SetInfo(CInventoryItem& pInvItem)
 
 	if (pAfContainer)
 	{
-		val = pAfContainer->GetContainerSize();
+		val = (float)pAfContainer->GetContainerSize();
 		if (!fis_zero(val))
 		{
 			m_artefacts_count->SetValue(val);
